app/requirements.txt
Flask>=2.3.0
flask-httpauth>=4.8.0
Werkzeug>=2.3.0
pymongo>=4.6.0
slack-sdk>=3.23.0
httpx>=0.25.0
schedule>=1.2.0
python-dotenv>=1.0.0
psutil>=5.9.0
Jinja2>=3.1.0
click>=8.1.0
itsdangerous>=2.1.0
MarkupSafe>=2.1.0
certifi>=2023.7.22
